Refactor TransitionSetCreator pt2
Avoid searching for transition views in both appearing and disappearing ViewControllers
