--- lib/android/app/src/main/java/com/reactnativenavigation/utils/View.kt
@@ -1,8 +1,0 @@
-package com.reactnativenavigation.utils
-
-import android.view.View
-import androidx.core.view.doOnLayout
-
-inline fun View.doIfMeasuredOrOnLayout(crossinline action: () -> Unit) {
-    if (width > 0 && height > 0) action() else doOnLayout { action() }
-}
