+++ lib/android/app/src/main/java/com/reactnativenavigation/views/element/TransitionSetCreator.kt
@@ -5,7 +5,8 @@ import com.reactnativenavigation.options.ElementTransitions
 import com.reactnativenavigation.options.NestedAnimationsOptions
 import com.reactnativenavigation.options.SharedElements
 import com.reactnativenavigation.viewcontrollers.viewcontroller.ViewController
-import com.reactnativenavigation.views.element.finder.ViewFinder
+import com.reactnativenavigation.views.element.finder.ExistingViewFinder
+import com.reactnativenavigation.views.element.finder.OptimisticViewFinder
 
 class TransitionSetCreator {
     fun create(animation: NestedAnimationsOptions, fromScreen: ViewController<*>, toScreen: ViewController<*>, onAnimatorsCreated: (TransitionSet) -> Unit) {
@@ -23,11 +24,11 @@ class TransitionSetCreator {
     private fun createSharedElementTransitions(fromScreen: ViewController<*>, toScreen: ViewController<*>, transitionSet: TransitionSet, sharedElements: SharedElements, elementTransitions: ElementTransitions, onTransitionCreated: (TransitionSet) -> Unit) {
         for (transitionOptions in sharedElements.get()) {
             val transition = SharedElementTransition(toScreen, transitionOptions)
-            ViewFinder().find(fromScreen, transition.fromId) {
+            OptimisticViewFinder().find(fromScreen, transition.fromId) {
                 transition.from = it
                 reportTransitionCreated(transitionSet, sharedElements, elementTransitions, transition, onTransitionCreated)
             }
-            ViewFinder().find(toScreen, transition.toId) {
+            OptimisticViewFinder().find(toScreen, transition.toId) {
                 transition.to = it
                 reportTransitionCreated(transitionSet, sharedElements, elementTransitions, transition, onTransitionCreated)
             }
@@ -44,11 +45,12 @@ class TransitionSetCreator {
     private fun createElementTransitions(fromScreen: ViewController<*>, toScreen: ViewController<*>, transitionSet: TransitionSet, sharedElements: SharedElements, elementTransitions: ElementTransitions, onAnimatorsCreated: (TransitionSet) -> Unit) {
         for (transitionOptions in elementTransitions.transitions) {
             val transition = ElementTransition(transitionOptions)
-            ViewFinder().find(fromScreen, transition.id) {
+            ExistingViewFinder().find(fromScreen, transition.id) {
                 transition.viewController = fromScreen
                 reportTransitionCreated(transitionSet, sharedElements, elementTransitions, transition, it, onAnimatorsCreated)
             }
-            ViewFinder().find(toScreen, transition.id) {
+            if (transition.isValid()) continue
+            OptimisticViewFinder().find(toScreen, transition.id) {
                 transition.viewController = toScreen
                 reportTransitionCreated(transitionSet, sharedElements, elementTransitions, transition, it, onAnimatorsCreated)
             }
