+++ lib/android/app/src/main/java/com/reactnativenavigation/views/element/finder/ViewFinder.kt
@@ -1,19 +1,8 @@
 package com.reactnativenavigation.views.element.finder
 
 import android.view.View
-import androidx.core.view.doOnLayout
-import com.facebook.react.uimanager.util.ReactFindViewUtil
 import com.reactnativenavigation.viewcontrollers.viewcontroller.ViewController
 
-class ViewFinder {
-    fun find(root: ViewController<*>, nativeId: String, onViewFound: (View) -> Unit) {
-        ReactFindViewUtil.findView(root.view, nativeId)
-                ?.let { onViewFound(it) }
-                ?: run {
-                    ReactFindViewUtil.findView(root.view, object : ReactFindViewUtil.OnViewFoundListener {
-                        override fun getNativeId() = nativeId
-                        override fun onViewFound(view: View) = view.doOnLayout { onViewFound(view) }
-                    })
-                }
-    }
+interface ViewFinder {
+    fun find(root: ViewController<*>, nativeId: String, onViewFound: (View) -> Unit)
 }
